Backend/requirements.txt
# HuggingFace API and LLM dependencies
transformers>=4.46.3          # HuggingFace transformers library for model loading and inference
torch>=1.13.1                 # PyTorch for running the models
huggingface_hub>=0.13.2       # HuggingFace Hub for model downloading and authentication
python-dotenv>=1.0.1          # To load environment variables from .env (for HuggingFace token)
accelerate>=0.17.1            # For efficient distributed training/inference

# Paper2Code dependencies (from PaperCoder)
openai>=1.65.4                # OpenAI API for GPT models
vllm>=0.6.4.post1             # vLLM for open-source model inference
tiktoken>=0.9.0               # Token counting for OpenAI models

# ASGI server for API serving
uvicorn==0.18.2

# Gradio 
gradio==3.3.0                 

# Other utilities
datasets>=2.10.1              # HuggingFace dataset library for managing datasets
requests>=2.28.1              # For handling HTTP requests
tqdm                          # Progress bars for training
